+++ src/policy/index.ts
@@ -4,3 +4,4 @@ export * from "./file-size";
 export * from "./patterns";
 export * from "./evaluate";
 export * from "./apex";
+export * from "./guards";